requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
pymongo==3.12.0
pydantic==2.12.2
pydantic-core==2.41.4
email-validator==2.2.0
bcrypt==4.2.1
python-multipart==0.0.20
jinja2==3.1.5
certifi
aiosmtplib==3.0.1
Pillow==12.0.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

